refactor(storage): rename db to boards-db
Self-mirror the module's own exports — getBoardsDB, BoardsDB, and the
"aac-boards-db" database name — rather than the bare "db".
