+++ src/features/board/board-set/delete-dialog.tsx
@@ -5,7 +5,7 @@ import DialogContent from "@mui/material/DialogContent";
 import DialogContentText from "@mui/material/DialogContentText";
 import DialogTitle from "@mui/material/DialogTitle";
 import { m } from "@paraglide/messages.js";
-import type { BoardSetRecord } from "../storage/db";
+import type { BoardSetRecord } from "../storage/boards-db";
 
 export interface BoardSetDeleteDialogProps {
   boardSet: BoardSetRecord | null;
