+++ src/features/board/board-set/info-dialog.tsx
@@ -8,7 +8,7 @@ import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
 import { m } from "@paraglide/messages.js";
 import { getEnglishLanguageName } from "@shared/utils/locale";
-import type { BoardSetRecord } from "../storage/db";
+import type { BoardSetRecord } from "../storage/boards-db";
 
 export interface BoardSetInfoDialogProps {
   boardSet: BoardSetRecord | null;
