+++ src/features/board/index.ts
@@ -11,10 +11,10 @@ export {
 } from "./navigation/board-paths";
 export { BoardNotFoundError, hydrateBoard } from "./storage/board-hydration";
 export { deleteBoardSet, getBoardSets } from "./storage/board-sets-store";
-export { getBoardSet } from "./storage/db";
+export { getBoardSet } from "./storage/boards-db";
 export { useBoardSets } from "./storage/use-board-sets";
 export { resolveTranslatedBoard } from "./translation/resolve-translated-board";
 
 export type { BoardRouteParams } from "./navigation/use-board-navigation";
-export type { BoardSetRecord } from "./storage/db";
+export type { BoardSetRecord } from "./storage/boards-db";
 export type { Board } from "./types";
