+++ src/features/board/storage/board-hydration.test.ts
@@ -2,7 +2,7 @@ import type { OBFBoard } from "@shayc/open-board-format";
 import { beforeEach, describe, expect, test } from "vitest";
 import { BoardNotFoundError, hydrateBoard } from "./board-hydration";
 import { invalidateBoardSets } from "./board-sets-store";
-import { putAssets, putBoards, upsertBoardSet } from "./db";
+import { putAssets, putBoards, upsertBoardSet } from "./boards-db";
 import { resetBoardsDB } from "./test-helpers";
 
 const SET_ID = "loader-test-set";
