+++ src/features/board/storage/board-hydration.ts
@@ -5,7 +5,7 @@ import {
 import type { OBFBoard, OBFMedia } from "@shayc/open-board-format";
 import { obfToBoard } from "../obf/obf-to-board";
 import type { Board } from "../types";
-import { getAssetBlob, getBoard } from "./db";
+import { getAssetBlob, getBoard } from "./boards-db";
 
 export class BoardNotFoundError extends Error {
   constructor(setId: string, boardId: string) {
