+++ src/features/board/storage/test-helpers.ts
@@ -1,5 +1,5 @@
 import { invalidateBoardSets } from "./board-sets-store";
-import { getBoardsDB, upsertBoardSet, type BoardSetRecord } from "./db";
+import { getBoardsDB, upsertBoardSet, type BoardSetRecord } from "./boards-db";
 
 const STORE_NAMES = ["boardSets", "boards", "assets"] as const;
