+++ src/features/board/translation/resolve-translated-board.test.ts
@@ -3,7 +3,7 @@ import {
   stubTranslator,
 } from "@shared/testing/built-in-ai";
 import { beforeEach, describe, expect, test } from "vitest";
-import { putBoards } from "../storage/db";
+import { putBoards } from "../storage/boards-db";
 import { resetBoardsDB, seedBoardSets } from "../storage/test-helpers";
 import type { Board } from "../types";
 import { resolveTranslatedBoard } from "./resolve-translated-board";
